name changes for subplugin
