+++ db/subplugins.php
@@ -24,4 +24,4 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$subplugins = (array) json_decode(file_get_contents(__DIR__ . "/subplugins.json"))->plugintypes;
+$subplugins = ['supportedmodules' => 'blocks/townsquare/supportedmodules'];
