--- supportedmodules/ts_moodleoverflow/db/access.php
@@ -1,45 +1,0 @@
-<?php
-// This file is part of Moodle - https://moodle.org/
-//
-// Moodle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Moodle is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
-
-/**
- * Plugin capabilities for the block_ts_moodleoverfloow subplugin.
- *
- * @package   block_ts_moodleoverflow
- * @copyright 2024 Tamaro Walter
- * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-defined('MOODLE_INTERNAL') || die();
-
-$capabilities = [
-    'block/townsquare/supportedmodules/block_ts_moodleoverflow:myaddinstance' => [
-        'captype' => 'write',
-        'contextlevel' => CONTEXT_SYSTEM,
-        'archetypes' => [
-            'user' => CAP_PROHIBIT,
-        ],
-        'clonepermissionsfrom' => 'moodle/my:manageblocks',
-    ],
-
-    'block/townsquare/supportedmodules/block_ts_moodleoverflow/towwnsquare:addinstance' => [
-        'riskbitmask' => RISK_SPAM | RISK_XSS,
-        'captype' => 'write',
-        'contextlevel' => CONTEXT_BLOCK,
-        'archetypes' => [
-            'user' => CAP_PROHIBIT,
-        ],
-        'clonepermissionsfrom' => 'moodle/site:manageblocks',
-    ],
-];
