+++ supportedmodules/tsmoodleoverflow/version.php
@@ -14,30 +14,18 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
 
-defined('MOODLE_INTERNAL') || die;
-
-use block_townsquare\townsquaresupport\townsquaresupportinterface;
-
-global $CFG;
-require_once($CFG->dirroot . '/blocks/townsquare/supportedmodules/block_ts_moodleoverflow/locallib.php');
-
 /**
- * Plugin strings are defined here.
+ * Plugin version and other meta-data are defined here.
  *
- * @package     block_townsquare
+ * @package     supportedmodules_tsmoodleoverflow
  * @copyright   2023 Tamaro Walter
  * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class ts__moodleoverflow implements townsquaresupportinterface {
 
-    /**
-     * Function from the interface.
-     * @return array
-     */
-    public function get_events(): array {
-        $courses = townsquare_get_courses();
-        $timestart = townsquare_get_timestart();
-        return ts_moodleoverflow_get_events($courses, $timestart);
-    }
-
-}
+defined('MOODLE_INTERNAL') || die();
+$plugin->component = 'supportedmodules_tsmoodleoverflow';
+$plugin->dependencies = ['block_townsquare' => ANY_VERSION];
+$plugin->release = '0.1.0';
+$plugin->version = 2024011503;
+$plugin->requires = 2022041900;
+$plugin->maturity = MATURITY_ALPHA;
